more tests
git-svn-id: https://zenmagick.svn.sourceforge.net/svnroot/zenmagick/trunk@1557 2e913b60-282b-11dc-bd19-815372732833

+++ zenmagick/plugins/request/zm_tests/tests/service/TestZMAccounts.php
@@ -64,6 +64,11 @@ class TestZMAccounts extends UnitTestCase {
             $ids[] = $result['accountId'];
         }
 
+        if (0 == count($ids)) {
+            // nothing to do
+            return;
+        }
+
         $sql = 'DELETE FROM '.TABLE_CUSTOMERS_INFO.' WHERE customers_info_id IN (:accountId)';
         $results = ZMRuntime::getDatabase()->update($sql, array('accountId' => $ids), TABLE_CUSTOMERS_INFO);
 
@@ -131,6 +136,30 @@ class TestZMAccounts extends UnitTestCase {
 
         $accounts = ZMAccounts::instance()->getAccountsForEmailAddress($account2->getEmail());
         $this->assertTrue(2 == count($accounts));
+    }
+
+    /**
+     * Test products subscriptions.
+     */
+    public function testProductSubscriptions() {
+        // delete previous subscriptions
+        $sql = "DELETE from " . TABLE_PRODUCTS_NOTIFICATIONS . "
+                WHERE  customers_id = :accountId";
+        ZMRuntime::getDatabase()->update($sql, array('accountId' => 2), TABLE_PRODUCTS_NOTIFICATIONS);
+
+        $testProductIds = array(1, 4, 7);
+        // insert new
+        $sql = "INSERT into " . TABLE_PRODUCTS_NOTIFICATIONS . "
+                (products_id, customers_id) VALUES(:productId, :accountId)";
+        foreach ($testProductIds as $id) {
+            ZMRuntime::getDatabase()->update($sql, array('accountId' => 2, 'productId' => $id), TABLE_PRODUCTS_NOTIFICATIONS);
+        }
+
+        $subscribedProductIds = array_flip(ZMAccounts::instance()->getSubscribedProductIds(2));
+        foreach ($testProductIds as $id) {
+            $this->assertTrue(array_key_exists($id, $subscribedProductIds));
+        }
+
     }
 
 }
